+++ examples/better-auth/src/lib/auth-client.ts
@@ -1,3 +1,6 @@
 import { createAuthClient } from "better-auth/react";
+import { siwsClient } from "./siws";
 
-export const authClient = createAuthClient({});
+export const authClient = createAuthClient({
+  plugins: [siwsClient()],
+});
