+++ examples/better-auth/src/lib/auth.ts
@@ -1,3 +1,10 @@
 import { betterAuth } from "better-auth";
+import { siws } from "./siws";
 
-export const auth = betterAuth({});
+export const auth = betterAuth({
+  plugins: [
+    siws({
+      domain: "localhost:3000",
+    }),
+  ],
+});
